wip: support for random metric selection in single pod
