+++ base-template/docker/bcc-sidecar/programs/open_close/ebpf-collect.c
