+++ base-template/templates/_helpers.tpl
@@ -93,9 +93,9 @@ Iterations is not relevant for this one
 {{- end }}
 
 {{ define "chart.monitor" }}
-  {{ if .Values.experiment.monitor }}{{- $progs := .Values.experiment.monitor | splitList "|" }}{{- range $index, $prog := $progs }}
+  {{ if .Values.experiment.monitor }}{{- $progs := .Values.experiment.monitor | splitList "|" }}
   - image: "ghcr.io/converged-computing/bcc-sidecar:ubuntu2204"
-    name: bcc-monitor-{{ $prog }}
+    name: bcc-monitor
     runFlux: false
     pullAlways: true 
     securityContext:
@@ -113,8 +113,8 @@ Iterations is not relevant for this one
         path: /sys/kernel/debug
     commands:
       pre: echo "ulimit -l unlimited" >> /root/.bashrc
-    command: "ulimit -l unlimited && ulimit -l && {{ if $.Values.experiment.monitor_command }}{{ $.Values.experiment.monitor_command }}{{ else }}python3 /opt/programs/{{ $prog }}/run-ebpf-collect.py --start-indicator-file=/mnt/flux/start_ebpf_collection {{ if $.Values.experiment.monitor_debug }}--debug{{ end }} --json {{ if $.Values.experiment.monitor_target }}--include-pattern={{ $.Values.experiment.monitor_target }}{{ end }} --stop-indicator-file=/mnt/flux/stop_ebpf_collection{{ end }}"{{ end }}
-{{- end }}{{- end }}
+    command: "ulimit -l unlimited && ulimit -l && {{ if .Values.experiment.monitor_command }}{{ .Values.experiment.monitor_command }}{{ else }}python3 /opt/programs/ebpf_collect.py --nodes {{ .Values.experiment.nodes }} --start-indicator-file=/mnt/flux/start_ebpf_collection {{ if .Values.experiment.monitor_debug }}--debug{{ end }} --json {{ if .Values.experiment.monitor_target }}--include-pattern={{ .Values.experiment.monitor_target }}{{ end }} --stop-indicator-file=/mnt/flux/stop_ebpf_collection{{ end }} {{- range $index, $prog := $progs }} -p {{ $prog }} {{- end -}}"{{ end }}
+{{- end }}
 
 {{/* Recording of application libraries 
 Currently only supported for single nodes and debian, requires proot.
